reorganized auth routes
