+++ src/routes/auth/2fa.js
@@ -1,8 +1,8 @@
 import { Router } from "express";
-import { signLogin, verify2FALogin } from "../utils/jwt.js";
-import { User } from "../models/user.js";
-import { generate2FASecret, verifyTOTP } from "../utils/totp.js";
-import { authenticate } from "../middleware/authentication.js"
+import { signLogin, verify2FALogin } from "../../utils/jwt.js";
+import { User } from "../../models/user.js";
+import { generate2FASecret, verifyTOTP } from "../../utils/totp.js";
+import { authenticate } from "../../middleware/authentication.js"
 
 const router = Router();
