+++ src/routes/auth/verifyEmail.js
@@ -1,8 +1,8 @@
 import { Router } from "express";
-import { User } from "../models/user.js";
-import { env } from "../config/env.js";
-import { sendMail } from "../config/nodemailer.js";
-import { verifyEmail, signEmailVerification } from "../utils/jwt.js";
+import { User } from "../../models/user.js";
+import { env } from "../../config/env.js";
+import { sendMail } from "../../config/nodemailer.js";
+import { verifyEmail, signEmailVerification } from "../../utils/jwt.js";
 
 const router = Router();
